pyproject.toml
[project]
name = "mini-pharmamind"
version = "0.1.0"
description = "Mini-PharmaMind: A lightweight proof-of-concept multi-agent system for pharmaceutical research (limited version)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "MHLAINE Khireddine", email = "mhalaine.khireddine.chimie@gmail.com"}
]
keywords = [
    "drug-discovery",
    "pharmaceutical-research",
    "multi-agent-system",
    "ai",
    "chemoinformatics",
    "bioinformatics"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Chemistry",
]


dependencies = [
    "autogen_agentchat==0.7.4",
    "autogen_core==0.7.4",
    "autogen_ext[mcp, openai]==0.7.4",
    "biopython==1.85",
    "chainlit==2.8.2",
    "chembl_webresource_client==0.10.9",
    "fpdf==1.7.2",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "prisma==0.15.0",
    "asyncpg==0.30.0",
    "mlflow>=3.8.1",
]


[project.optional-dependencies]

dev = [
    "black == 25.9.0",
    "flake8 == 7.3.0",
    "mypy == 1.18.2",
    "pylint == 4.0.2",
    "pytest == 8.4.2",
    "pytest-asyncio == 1.2.0"
]

